Add tests and organise files
